chore(core): comprehensive verification and cleanup
- Applied strict clippy fixes across workspace for brain and storage modules
- Reformatted codebase with cargo fmt
- Resolved unused imports in network flow control
- Verified all 60+ tests pass with full feature set

+++ src/ui/renderer.rs
@@ -161,8 +161,8 @@ impl<'a> Widget for WorldWidget<'a> {
             }
         }
 
-        let map_w = self.snapshot.terrain.width as u16;
-        let map_h = self.snapshot.terrain.height as u16;
+        let map_w = self.snapshot.terrain.width;
+        let map_h = self.snapshot.terrain.height;
 
         let start_x = 0;
         let end_x = inner.width.min(map_w);
